+++ main-server/src/http_server/routes/v6/rp.js
@@ -82,7 +82,6 @@ router.post(
         },
         {
           synchronous: false,
-          validateIdentifier: false,
           apiVersion,
           ndidMemberAppType,
           ndidMemberAppVersion,